Ajourne les drapeaux de compilation, correction des erreurs de compilation y découlant.

+++ coeur/danjo/controles/controle_couleur.cc
@@ -41,7 +41,7 @@ ControleCouleur::ControleCouleur(QWidget *parent)
 			this, &ControleCouleur::ajourne_couleur);
 
 	const auto &metriques = this->fontMetrics();
-	setFixedHeight(metriques.height() * 1.5f);
+	setFixedHeight(static_cast<int>(static_cast<float>(metriques.height()) * 1.5f));
 	setFixedWidth(metriques.width("#000000"));
 }
